Improvements and cleaning
