+++ .vscode/c_cpp_properties.json
@@ -477,7 +477,7 @@
             ],
             "cStandard": "gnu99",
             "cppStandard": "gnu++11",
-            "compilerPath": "C:/Users/Naervern/.platformio/packages/toolchain-xtensa-esp32s2@8.4.0+2021r2-patch5/bin/xtensa-esp32s2-elf-gcc.exe",
+            "compilerPath": "C:\\Users\\Naervern\\.platformio\\packages\\toolchain-xtensa-esp32s2@8.4.0+2021r2-patch5\\bin\\xtensa-esp32s2-elf-gcc.exe",
             "compilerArgs": [
                 "-mlongcalls",
                 ""
