+++ .vscode/launch.json
@@ -1,56 +1,63 @@
+// AUTOMATICALLY GENERATED FILE. PLEASE DO NOT MODIFY IT MANUALLY
+//
+// PlatformIO Debugging Solution
+//
+// Documentation: https://docs.platformio.org/en/latest/plus/debugging.html
+// Configuration: https://docs.platformio.org/en/latest/projectconf/sections/env/options/debug/index.html
+
 {
-  "version": "0.2.0",
-  "configurations": [
-    {
-      "type": "platformio-debug",
-      "request": "launch",
-      "name": "PIO Debug",
-      "executable": "c:/Users/Naervern/Documents/GitHub/meteo-nano/.pio/build/lolin_s2_mini/firmware.elf",
-      "projectEnvName": "lolin_s2_mini",
-      "toolchainBinDir": "C:/Users/Naervern/.platformio/packages/toolchain-xtensa-esp32s2@8.4.0+2021r2-patch5/bin",
-      "internalConsoleOptions": "openOnSessionStart",
-      "preLaunchTask": {
-        "type": "PlatformIO",
-        "task": "Pre-Debug"
-      }
-    },
-    {
-      "type": "platformio-debug",
-      "request": "launch",
-      "name": "PIO Debug (skip Pre-Debug)",
-      "executable": "c:/Users/Naervern/Documents/GitHub/meteo-nano/.pio/build/lolin_s2_mini/firmware.elf",
-      "projectEnvName": "lolin_s2_mini",
-      "toolchainBinDir": "C:/Users/Naervern/.platformio/packages/toolchain-xtensa-esp32s2@8.4.0+2021r2-patch5/bin",
-      "internalConsoleOptions": "openOnSessionStart"
-    },
-    {
-      "type": "platformio-debug",
-      "request": "launch",
-      "name": "PIO Debug (without uploading)",
-      "executable": "c:/Users/Naervern/Documents/GitHub/meteo-nano/.pio/build/lolin_s2_mini/firmware.elf",
-      "projectEnvName": "lolin_s2_mini",
-      "toolchainBinDir": "C:/Users/Naervern/.platformio/packages/toolchain-xtensa-esp32s2@8.4.0+2021r2-patch5/bin",
-      "internalConsoleOptions": "openOnSessionStart",
-      "loadMode": "manual"
-    },
-    {
-      "name": "C/C++ Runner: Debug Session",
-      "type": "cppdbg",
-      "request": "launch",
-      "args": [],
-      "stopAtEntry": false,
-      "externalConsole": true,
-      "cwd": "c:/Users/Naervern/Documents/GitHub/meteo-nano/lib/BME280_I2C/src",
-      "program": "c:/Users/Naervern/Documents/GitHub/meteo-nano/lib/BME280_I2C/src/build/Debug/outDebug",
-      "MIMode": "gdb",
-      "miDebuggerPath": "gdb",
-      "setupCommands": [
+    "version": "0.2.0",
+    "configurations": [
         {
-          "description": "Enable pretty-printing for gdb",
-          "text": "-enable-pretty-printing",
-          "ignoreFailures": true
+            "name": "C/C++ Runner: Debug Session",
+            "type": "cppdbg",
+            "request": "launch",
+            "args": [],
+            "stopAtEntry": false,
+            "externalConsole": true,
+            "cwd": "c:/Users/Naervern/Documents/GitHub/meteo-nano",
+            "program": "c:/Users/Naervern/Documents/GitHub/meteo-nano/build/Debug/outDebug",
+            "MIMode": "gdb",
+            "miDebuggerPath": "gdb",
+            "setupCommands": [
+                {
+                    "description": "Enable pretty-printing for gdb",
+                    "text": "-enable-pretty-printing",
+                    "ignoreFailures": true
+                }
+            ]
+        },
+        {
+            "type": "platformio-debug",
+            "request": "launch",
+            "name": "PIO Debug",
+            "executable": "c:/Users/Naervern/Documents/GitHub/meteo-nano/.pio/build/lolin_s2_mini/firmware.elf",
+            "projectEnvName": "lolin_s2_mini",
+            "toolchainBinDir": "C:/Users/Naervern/.platformio/packages/toolchain-xtensa-esp32s2@8.4.0+2021r2-patch5/bin",
+            "internalConsoleOptions": "openOnSessionStart",
+            "preLaunchTask": {
+                "type": "PlatformIO",
+                "task": "Pre-Debug"
+            }
+        },
+        {
+            "type": "platformio-debug",
+            "request": "launch",
+            "name": "PIO Debug (skip Pre-Debug)",
+            "executable": "c:/Users/Naervern/Documents/GitHub/meteo-nano/.pio/build/lolin_s2_mini/firmware.elf",
+            "projectEnvName": "lolin_s2_mini",
+            "toolchainBinDir": "C:/Users/Naervern/.platformio/packages/toolchain-xtensa-esp32s2@8.4.0+2021r2-patch5/bin",
+            "internalConsoleOptions": "openOnSessionStart"
+        },
+        {
+            "type": "platformio-debug",
+            "request": "launch",
+            "name": "PIO Debug (without uploading)",
+            "executable": "c:/Users/Naervern/Documents/GitHub/meteo-nano/.pio/build/lolin_s2_mini/firmware.elf",
+            "projectEnvName": "lolin_s2_mini",
+            "toolchainBinDir": "C:/Users/Naervern/.platformio/packages/toolchain-xtensa-esp32s2@8.4.0+2021r2-patch5/bin",
+            "internalConsoleOptions": "openOnSessionStart",
+            "loadMode": "manual"
         }
-      ]
-    }
-  ]
+    ]
 }
